Extract base response model for secret
